+++ ogre/include/ignition/rendering/ogre/OgreScene.hh
@@ -133,6 +133,10 @@ namespace ignition
                      const std::string &_name);
 
       // Documentation inherited
+      protected: virtual MarkerPtr CreateMarkerImpl(unsigned int _id,
+                     const std::string &_name);
+
+      // Documentation inherited
       protected: virtual TextPtr CreateTextImpl(unsigned int _id,
                      const std::string &_name);